misc: restructure extras and utilities
