--- docs/api/collections/validators.md
@@ -1,1 +1,0 @@
-::: opensourceleg.collections.validators
