+++ docs/api/extras/collections/validators.md
@@ -1,0 +1,1 @@
+::: opensourceleg.extras.collections.validators
