+++ docs/api/extras/safety/safety.md
@@ -1,0 +1,1 @@
+::: opensourceleg.extras.safety
