+++ docs/api/math/math.md
@@ -1,0 +1,1 @@
+::: opensourceleg.math.math
