--- docs/api/safety.md
@@ -1,1 +1,0 @@
-::: opensourceleg.safety
