--- docs/api/time.md
@@ -1,1 +1,0 @@
-::: opensourceleg.time
