+++ docs/api/utilities/softrealtimeloop.md
@@ -1,0 +1,1 @@
+::: opensourceleg.utilities.softrealtimeloop
