--- opensourceleg/units/__init__.py
@@ -1,10 +1,0 @@
-from .units import *  # noqa: F403
-
-"""
-Units module for opensourceleg library.
-
-Module Overview:
-
-This module provides classes and functions for managing units in the opensourceleg library.
-
-"""
