+++ opensourceleg/utilities/units.py
@@ -16,19 +16,19 @@ Usage:
 """
 
 __all__ = [
-    "Force",
-    "Torque",
-    "Stiffness",
-    "Damping",
-    "Length",
-    "Position",
-    "Mass",
-    "Velocity",
     "Acceleration",
     "Current",
+    "Damping",
+    "Force",
+    "Length",
+    "Mass",
+    "Position",
+    "Stiffness",
+    "Torque",
+    "Velocity",
     "Voltage",
-    "convert_to_default",
     "convert_from_default",
+    "convert_to_default",
 ]
