+++ tutorials/actuators/dephy/commanding_voltage.py
@@ -3,7 +3,7 @@ import time
 from opensourceleg.actuators.base import CONTROL_MODES
 from opensourceleg.actuators.dephy import DephyActuator
 from opensourceleg.logging.logger import Logger
-from opensourceleg.time import SoftRealtimeLoop
+from opensourceleg.utilities import SoftRealtimeLoop
 
 FREQUENCY = 1000
 TIME_TO_STEP = 1.0
