+++ tutorials/utilities/using_iterators.py
@@ -1,4 +1,4 @@
-from opensourceleg.time import SoftRealtimeLoop
+from opensourceleg.utilities import SoftRealtimeLoop
 
 
 def main():
